+++ server/src/authfort/providers/__init__.py
@@ -1,6 +1,7 @@
 """AuthFort OAuth providers."""
 
 from authfort.providers.base import OAuthProvider, OAuthUserInfo
+from authfort.providers.generic import GenericOAuthProvider, GenericOIDCProvider
 from authfort.providers.github import GitHubProvider
 from authfort.providers.google import GoogleProvider
 
@@ -9,4 +10,6 @@ __all__ = [
     "OAuthUserInfo",
     "GoogleProvider",
     "GitHubProvider",
+    "GenericOAuthProvider",
+    "GenericOIDCProvider",
 ]